🔧 Config. PyUp to run monthly on development (#45)
While roman-numerals is still under development, the weekly updates
affect the development setup, not the actual security of the package
itself. As such, to save myself time, I'm switching the PyUp bot to open
PRs on a monthly basis.

This also adds an emoji prefix to the PRs. For reasons.